src/requirements.txt
# https://developer.apple.com/metal/tensorflow-plugin/
keras~=2.9
psycopg2-binary~=2.9.3
tensorflow~=2.12.0; sys_platform != 'darwin'
tensorflow-macos~=2.9.0; sys_platform == 'darwin'
tensorflow-metal==0.5.0; sys_platform == 'darwin'
numpy~=1.22
Flask~=2.0.1
pandas~=1.4.1
matplotlib~=3.5.1
plotly~=5.5.0
joblib~=1.1.0
scikit-learn~=1.0.2
# gunicorn

pip~=22.1.2
attrs~=22.1.0
zipp~=3.7.0
toml~=0.10.2
wheel~=0.37.1
Pillow~=9.0.1
Jinja2~=3.0.1
setuptools~=63.1.0
kafka-python~=2.0.2
